use float for mask

+++ utility.py
@@ -33,7 +33,7 @@ def rle_decode(mask_rle, shape):
   starts, lengths = [np.asarray(x, dtype=int) for x in (s[0:][::2], s[1:][::2])]
   starts -= 1
   ends = starts + lengths
-  img = np.zeros(shape[0]*shape[1], dtype=np.uint8)
+  img = np.zeros(shape[0]*shape[1], dtype=np.float)
   for lo, hi in zip(starts, ends):
       img[lo:hi] = 1
   return img.reshape(shape[1], shape[0]).T
